requirements.txt
openai>=1.40.0
azure-identity>=1.15.0
python-dotenv>=1.0.0

# Evaluation (Azure AI Foundry cloud evaluation)
azure-ai-evaluation>=1.0.0
azure-ai-projects>=2.0.0b3

# Agent Framework (optional, for agent-based evaluation)
azure-ai-agents>=1.2.0b5
